Make velocity_smoother test work with 'colcon test'.
It seems that 'colcon test' has a different execution path
than 'ros2 test', so we can't rely on the launch_ros_node
being available.  Make sure to create our own node, and then
use that for the subscriptions.  This also necessitated a
change in how the checking loop was done.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>